+++ server/Recruit-Api/src/main/java/com/econovation/recruit/api/interviewer/service/InterviewersService.java
@@ -3,11 +3,11 @@ package com.econovation.recruit.api.interviewer.service;
 import com.econovation.recruit.api.interviewer.helper.IdpHelper;
 import com.econovation.recruit.api.interviewer.usecase.InterviewerUseCase;
 import com.econovation.recruitdomain.domains.dto.InterviewerCreateDto;
-import com.econovation.recruitdomain.domains.dto.InterviewerResponse;
 import com.econovation.recruitdomain.domains.interviewer.domain.Interviewer;
 import com.econovation.recruitdomain.domains.interviewer.domain.Role;
 import com.econovation.recruitdomain.out.InterviewerLoadPort;
 import com.econovation.recruitdomain.out.InterviewerRecordPort;
+import com.econovation.recruitinfrastructure.idp.dto.InterviewerResponse;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -22,7 +22,7 @@ public class InterviewersService implements InterviewerUseCase {
     private final IdpHelper idpHelper;
 
     @Override
-    public List<Interviewer> createInterviewers(List<Integer> idpIds) {
+    public List<Interviewer> createInterviewers(List<Long> idpIds) {
         List<InterviewerResponse> interviewers = idpHelper.createInterviewers(idpIds);
         return interviewerRecordPort.saveAll(
                 interviewers.stream().map(InterviewerCreateDto::from).collect(Collectors.toList()));
